refactor: move storage logic to storage.*
